Put list of boolean ops in bullet list
I tried putting them in a code block by indenting them 4 spaces like you do for examples, but that didn't seem to mean anything to the markup parser and it just ran the lines together in into one line like it would for normal text.  So now I'm putting int in a bullet list instead.