Update volumes.html
SET VOLUME("0"):NAME TO "newname". This code always throws an error "cannot find volume 0". However without the quotes it works as expected. I just removed the quotes so it now shows SET VOLUME(0):NAME TO "newname". It might also be helpful to change the name of volume 1 instead of the archive, but that isn't an error.