CREATEORBIT(...) had velocity and position args reversed
The documentation for the second form of CREATEORBIT() contained an error. The example given was correct, but the synopsis and description immediately above it were reversed. This commit fixes that.

I have tested this using the latest version of kOS in CKAN (1.4.0) and I can confirm that code following the updated documentation is working.